Restructure the tests folder
